requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
python-multipart==0.0.5
pika==1.2.0
requests==2.27.1
python-dotenv==0.19.2